The basic sensor triggering a job is working!
