--- visdag/jobs/__init__.py
@@ -1,1 +1,0 @@
-from .raw_jobs import raw_jobs
