--- visdag/jobs/raw_jobs/__init__.py
@@ -1,1 +1,0 @@
-from .raw_jobs import log_raw_json
